i doing questions scene

+++ commands/addImage.js
@@ -4,8 +4,13 @@ module.exports = (bot) => {
     bot.command('/add_images',ctx => {
         fsService.addImages()
             .then(res=>{
-                fsService.deleteImage(res.reject)
-                ctx.reply('this image is reject '+res.reject)
+                if(res.reject.length !== 0)
+                {
+                    fsService.deleteImage(res.reject)
+                    ctx.reply('this image is reject '+res.reject)
+                }else{
+                    ctx.reply('it is ok')
+                }
             })
     })
 }
